refactor function to class in dedicated file `cookie_string_builder.ts`
